+++ CHANGELOG.md
@@ -14,6 +14,7 @@ BUG FIXES:
     checking a box checksum. [GH-2716]
   - core: Better error message if your plugin state file becomes corrupt
     somehow. [GH-2694]
+  - core: Box add will fail early if the box already exists. [GH-2621]
   - hosts/bsd: Only run `nfsd checkexports` if there is an exports file.
     [GH-2714]
   - commands/plugin: Fix exception that could happen rarely when installing